processing/ingestion-service/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pandas==2.0.1
numpy==1.24.3
python-multipart==0.0.6
boto3==1.26.129
pydantic==1.10.7
sqlalchemy==2.0.12
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytest==7.3.1
httpx==0.24.0
